refactor Location component
